+++ client/src/api/search.js
@@ -43,8 +43,12 @@ export async function instantSearch (query) {
   return await get(`/search/all?q=${encodeURIComponent(query)}`)
 }
 
-export async function searchUser (query) {
-  return await get(`/search/user?q=${encodeURIComponent(query)}`)
+export async function searchUser (query, regionId = null) {
+  let path = `/search/user?q=${encodeURIComponent(query)}`
+  if (regionId !== null) {
+    path += `&regionId=${regionId}`
+  }
+  return await get(path)
 }
 
 export async function instantSearchIndex () {