+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 - Display same-day pickups when confirming to sign into a pickup slot !1827 @ChrisOelmueller
 - Show membership in profile for workgroups !1988 !1996 @chriswalg
 - IT-Support Admins can delete bananas upon request over it@foodsharing.network. !2002 @fs_k
-- Added a user search field to the store management panel !2007 @alex.simm
+- Added a user search field to the store management panel !2007 !2033 @alex.simm
 - New calendar API including token management #80 !1719 !2029 @alex.simm 
 
 ## Changes